#22: members serializer
